mathdroid/res/layout-land/main.xml: add a landscape layout without a keypad (on the not necessarily valid assumption that you're using a G1, and "landscape" means you have the hardware keyboard out).
mathdroid/res/layout/main.xml: follow the landscape UI's lead in putting the history above the EditText. It makes more sense that way round as long as we're appending to the history.

mathdroid/src/org/jessies/mathdroid/Mathdroid.java: the code needs to cope with the fact that buttons aren't guaranteed to exist in every layout.


git-svn-id: http://enh.googlecode.com/svn/trunk@223 825dafb7-1f56-0410-864d-61665d7775f4