mathdroid/src/org/jessies/mathdroid/HistoryAdapter.java: mathdroid/src/org/jessies/mathdroid/HistoryItem.java: mathdroid/src/org/jessies/mathdroid/HistoryItemView.java: everything we need to switch from a TextView for the history to using a ListView instead. Both writing and using the new ListView version, it's obvious that this is more natural for Android.
mathdroid/res/values/styles.xml: a style for the questions in the history; the answers get the default style.

mathdroid/res/layout/text_area.xml: switch to a ListView.

mathdroid/src/org/jessies/mathdroid/Mathdroid.java: switch to a ListView, add ceil, floor, and rand, and add left and right cursors to the on-screen keypad. This is useful if you have a device without a trackball. (Though it's a waste of space if you do have a trackball.)

mathdroid/res/layout/button_area.xml: changes to the on-screen keypad layout.


git-svn-id: http://enh.googlecode.com/svn/trunk@322 825dafb7-1f56-0410-864d-61665d7775f4